Print the meaning of the leap second field if it's non-zero
The meaning comes from RFC 5905 and is important because one of the
reasons is 'leap second state is unknown because the remote clock is
probably not synchronized', which causes a validity problem that is
reported as 'invalid leap second'. This way, we sort of get an
explanation of that, as opposed to a mystery.

(We were already printing the numeric value.)